requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.1
exceptiongroup==1.3.0
idna==3.10
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.1
pytest==8.4.0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
tabulate==0.9.0
tomli==2.2.1
typing_extensions==4.14.0
urllib3==2.4.0
